+++ package.json
@@ -15,6 +15,7 @@
     "react": "^16.13.0",
     "react-bootstrap": "^1.0.0-beta.17",
     "react-dom": "^16.13.0",
+    "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
     "react-scripts-ts": "3.1.0",
     "socket.io": "^2.3.0",
