+++ src/index.tsx
@@ -1,14 +1,12 @@
+/* tslint:disable */
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import App from './App';
 import './index.css';
-import { BrowserRouter as Router } from 'react-router-dom';
 import registerServiceWorker from './registerServiceWorker';
 
 ReactDOM.render(
-  <Router>
-    <App />
-  </Router>,
+    <App />,
   document.getElementById("root") as HTMLElement
 );
 registerServiceWorker();
